EclThermalLawManager: move code to separate compile unit
